Change file structure, move LoadAddedChannels into data layer
also the gradle version is updated to make sure eveyr developer
can download the source code and get it compiled directly
